requirements.txt
streamlit==1.15.0
pandas==1.5.3
google-auth==2.16.0
google-api-python-client==2.76.0
google-auth-httplib2==0.1.0
google-api-core==2.9.1
Pillow==9.2.0
PyPDF2==2.11.1
reportlab==3.6.11
streamlit-drawable-canvas==0.3.0
